# Code/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CGDA Backend",
    "description": "Minimal backend service for the geographic analysis platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/runtime/config": {
      "patch": {
        "tags": [
          "runtime"
        ],
        "summary": "Update Runtime Config",
        "operationId": "update_runtime_config_runtime_config_patch",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeConfigUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeConfigUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Get Runtime Config",
        "description": "Return the current runtime configuration snapshot (defaults + DB overrides).",
        "operationId": "get_runtime_config_runtime_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Runtime Config Runtime Config Get"
                }
              }
            }
          }
        }
      }
    },
...
    "/frontend/commands": {
      "post": {
        "tags": [
          "frontend"
        ],
        "summary": "Submit Frontend Command",
        "description": "已移除：前端控制指令为占位实现，无实际副作用。",
        "operationId": "submit_frontend_command_frontend_commands_post",
        "deprecated": true,
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FrontendCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrontendCommandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/runtime/api-config/{provider}": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Get Provider Api Config",
        "description": "返回指定 Provider 的 API 配置状态。",
        "operationId": "get_provider_api_config_runtime_api_config__provider__get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Update Provider Api Config",
        "description": "已移除：运行时密钥/配置请改用 /config/api-keys* 与 /config/weather/providers*。",
        "operationId": "update_provider_api_config_runtime_api_config__provider__post",
        "deprecated": true,
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Config Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime/api-config/{provider}/best": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Get Best Available Api",
        "description": "返回指定数据类型中最佳可用的 API Provider。",
        "operationId": "get_best_available_api_runtime_api_config__provider__best_get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/runtime/tiles/cache/clear": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Clear Runtime Tile Cache",
        "description": "清空底图代理进程内缓存。",
        "operationId": "clear_runtime_tile_cache_runtime_tiles_cache_clear_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Clear Runtime Tile Cache Runtime Tiles Cache Clear Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/unified-tiles/{layer_id}/{z}/{x}/{y}": {
      "get": {
        "tags": [
          "tiles"
        ],
        "summary": "Get Unified Tile",
        "description": "获取底图栅格瓦片。",
        "operationId": "get_unified_tile_unified_tiles__layer_id___z___x___y__get",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          },
          {
            "name": "use_cache",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Use Cache"
            }
          },
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "客户端缓存 bust，不参与业务",
              "title": "T"
            },
            "description": "客户端缓存 bust，不参与业务"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/weather/tiles/{layer_id}/{z}/{x}/{y}": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Get Weather Tile",
        "description": "获取指定图层的标准 Web Mercator GeoJSON 瓦片。",
        "operationId": "get_weather_tile_weather_tiles__layer_id___z___x___y__get",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          },
          {
            "name": "hour",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 47,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "default": 0,
              "title": "Hour"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "T"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClientIdentity": {
        "properties": {
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "page": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "view_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "View Id"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "type": "object",
        "title": "ClientIdentity"
      },
...
      "FrontendCommandRequest": {
        "properties": {
          "command_type": {
            "$ref": "#/components/schemas/FrontendCommandType"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "client": {
            "$ref": "#/components/schemas/ClientIdentity"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "command_type"
        ],
        "title": "FrontendCommandRequest"
      },
      "FrontendCommandResponse": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "command_type": {
            "$ref": "#/components/schemas/FrontendCommandType"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "next_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Action"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "command_type",
          "created_at",
          "message"
        ],
        "title": "FrontendCommandResponse"
      },
      "FrontendCommandType": {
        "type": "string",
        "enum": [
          "preload",
          "clear_cache",
          "cleanup",
          "cancel_run",
          "reload_catalog",
          "custom"
        ],
        "title": "FrontendCommandType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuntimeConfigPatch": {
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/RuntimeConfigScope"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "title": "Value"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "key",
          "value"
        ],
        "title": "RuntimeConfigPatch"
      },
      "RuntimeConfigScope": {
        "type": "string",
        "enum": [
          "frontend",
          "backend",
          "provider",
          "workflow",
          "system"
        ],
        "title": "RuntimeConfigScope"
      },
      "RuntimeConfigUpdateRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RuntimeConfigPatch"
            },
            "type": "array",
            "title": "Items"
          },
          "client": {
            "$ref": "#/components/schemas/ClientIdentity"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "RuntimeConfigUpdateRequest"
      },
      "RuntimeConfigUpdateResponse": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "applied_count": {
            "type": "integer",
            "title": "Applied Count"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "config_snapshot": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Config Snapshot"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "updated_at",
          "applied_count",
          "message",
          "config_snapshot"
        ],
        "title": "RuntimeConfigUpdateResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}